proto select frame
